Module Plist_xml

type t = [

| `Bool of bool

| `Data of string

| `Date of float * float option
(timestamp, timezone)

| `Float of float

| `Int of int

| `String of string

| `Array of t list
Array

| `Dict of (string * t) list
Dictionary

]

Plist values.

val signals : t -> (Markup.signal, Markup.sync) Markup.stream

Convert a plist into XML signals.

exception Parse_error of string

Exception raised upon a parse error. The error message is purely informational and you should not pattern match on it, as it is subject to change.

module type IO = sig ... end

The module type of synchronous or asynchronous I/O.

module IO : IO with type s = Markup.sync and type 'a io = 'a

Synchronous I/O under the identity monad.

module type S = sig ... end

module Make : functor (IO : IO) -> S with type s = IO.s and type 'a io = 'a IO.io

Create a plist parser given an I/O implementation, such as synchronous I/O or Lwt.

include S with type S.s = Markup.sync and type 'a S.io = 'a

type s = Markup.sync

type 'a io = 'a

val plist_of_stream_exn : (Markup.content_signal, s) Markup.stream -> t io

Raises Parse_error upon failure.

val parse_exn : ?⁠report:(Markup.location -> Markup.Error.t -> unit io) -> ?⁠encoding:Markup.Encoding.t -> ?⁠namespace:(string -> string option) -> ?⁠entity:(string -> string option) -> ?⁠context:[< `Document | `Fragment ] -> (char, s) Markup.stream -> t io

Raises Parse_error upon failure. See documentation for Markup.parse_xml for labeled parameter information.